+++ packages/dapp/src/components/Wallet.tsx
@@ -1,7 +1,7 @@
 import { Box } from "@mui/material";
 import { LoadingButton } from "@mui/lab";
-import { walletStore, WalletStatusEnums } from "./walletStore";
-import { nftsStore } from "../nfts/nftsStore";
+import { walletStore, WalletStatusEnums } from "../features/wallet/walletStore";
+import { nftsStore } from "../features/nfts/nftsStore";
 
 const Wallet = () => {
   const connected = walletStore.status === WalletStatusEnums.CONNECTED;
