+++ packages/dapp/src/components/map/index.ts
@@ -1,0 +1,1 @@
+export * from "./Map";
