--- packages/dapp/src/features/nfts/geojson0.json
@@ -1,36 +1,0 @@
-{
-    "type": "FeatureCollection",
-    "features": [
-      {
-        "type": "Feature",
-        "properties": {},
-        "geometry": {
-          "type": "Polygon",
-          "coordinates": [
-            [
-              [
-                -429.4033813476562,
-                11.821653169925586
-              ],
-              [
-                -428.5629272460937,
-                11.821653169925586
-              ],
-              [
-                -428.5629272460937,
-                12.61487691594163
-              ],
-              [
-                -429.4033813476562,
-                12.61487691594163
-              ],
-              [
-                -429.4033813476562,
-                11.821653169925586
-              ]
-            ]
-          ]
-        }
-      }
-    ]
-  }
