--- packages/dapp/src/features/nfts/geojson1.json
@@ -1,118 +1,0 @@
-{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "type": "Polygon",
-        "coordinates": [
-          [
-            [
-              -428.8906744122505,
-              12.147418397582491
-            ],
-            [
-              -428.8907468318939,
-              12.147347599447487
-            ],
-            [
-              -428.8907213509083,
-              12.14723615790054
-            ],
-            [
-              -428.8905939459801,
-              12.147198136656193
-            ],
-            [
-              -428.89051884412766,
-              12.147280734524921
-            ],
-            [
-              -428.89055103063583,
-              12.147379065287602
-            ],
-            [
-              -428.8906744122505,
-              12.147418397582491
-            ]
-          ]
-        ]
-      }
-    },
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "type": "Polygon",
-        "coordinates": [
-          [
-            [
-              -428.8905443251133,
-              12.147381687440772
-            ],
-            [
-              -428.89051616191864,
-              12.147279423447841
-            ],
-            [
-              -428.89041021466255,
-              12.147224358204612
-            ],
-            [
-              -428.8903096318245,
-              12.147300400680368
-            ],
-            [
-              -428.8903257250786,
-              12.147409220047532
-            ],
-            [
-              -428.8904584944248,
-              12.147449863414236
-            ],
-            [
-              -428.8905443251133,
-              12.147381687440772
-            ]
-          ]
-        ]
-      }
-    },
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "type": "Polygon",
-        "coordinates": [
-          [
-            [
-              -428.8901272416115,
-              12.147367265597998
-            ],
-            [
-              -428.89017820358276,
-              12.147195514501217
-            ],
-            [
-              -428.8900186121464,
-              12.147116849839737
-            ],
-            [
-              -428.8899327814579,
-              12.147217802817746
-            ],
-            [
-              -428.8899743556976,
-              12.147334488679682
-            ],
-            [
-              -428.8901272416115,
-              12.147367265597998
-            ]
-          ]
-        ]
-      }
-    }
-  ]
-}
