--- packages/dapp/src/features/nfts/geojson2.json
@@ -1,44 +1,0 @@
-{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "type": "Polygon",
-        "coordinates": [
-          [
-            [
-              -428.93541499972343,
-              12.110363491120767
-            ],
-            [
-              -428.93546998500824,
-              12.110321530841098
-            ],
-            [
-              -428.9354766905307,
-              12.110244166558177
-            ],
-            [
-              -428.9354056119919,
-              12.110210073816223
-            ],
-            [
-              -428.935339897871,
-              12.110244166558177
-            ],
-            [
-              -428.93533587455744,
-              12.110329398394029
-            ],
-            [
-              -428.93541499972343,
-              12.110363491120767
-            ]
-          ]
-        ]
-      }
-    }
-  ]
-}
