+++ packages/dapp/src/features/wallet/walletStore.ts
@@ -2,14 +2,13 @@ import { makeAutoObservable } from "mobx";
 import { ethers } from "ethers";
 import Web3Modal from "web3modal";
 import WalletConnectProvider from "@walletconnect/web3-provider";
-// import { create, IPFSHTTPClient } from "ipfs-http-client";
+import { create, IPFSHTTPClient } from "ipfs-http-client";
 
 class WalletStore {
   address: string | null = null;
   balance: string | null = null;
   status: WalletStatusEnums = WalletStatusEnums.DISCONNECTED;
-  // ipfsClient: IPFSHTTPClient | null = null;
-  ipfsClient = true; // Mock ipfsClient for now. It will only be used on image storage
+  ipfsClient: IPFSHTTPClient | null = null;
   provider: any | null = null;
   private web3Modal: Web3Modal | null = null;
 
@@ -71,7 +70,7 @@ class WalletStore {
       this.balance = ethers.utils.formatEther(balance);
       this.provider = provider;
       this.web3Modal = web3Modal;
-      // this.ipfsClient = create(ipfsOptions);
+      this.ipfsClient = create(ipfsOptions);
       this.status = WalletStatusEnums.CONNECTED;
 
       console.log("Connected to wallet");
@@ -97,21 +96,21 @@ class WalletStore {
   };
 }
 
-// const ipfsOptions = {
-//   host: "ipfs.infura.io",
-//   port: 5001,
-//   protocol: "https",
-//   apiPath: "/api/v0",
-//   headers: {
-//     authorization:
-//       "Basic " +
-//       Buffer.from(
-//         process.env.REACT_APP_PROJECT_ID +
-//           ":" +
-//           process.env.REACT_APP_PROJECT_SECRET
-//       ).toString("base64"),
-//   },
-// };
+const ipfsOptions = {
+  host: "ipfs.infura.io",
+  port: 5001,
+  protocol: "https",
+  apiPath: "/api/v0",
+  headers: {
+    authorization:
+      "Basic " +
+      Buffer.from(
+        process.env.REACT_APP_PROJECT_ID +
+          ":" +
+          process.env.REACT_APP_PROJECT_SECRET
+      ).toString("base64"),
+  },
+};
 
 export enum WalletStatusEnums {
   DISCONNECTED,
